Litmus WebDAV tester requirements

+++ lib/redmine_dmsf/webdav/dmsf_resource.rb
@@ -209,6 +209,11 @@ module RedmineDmsf
             Conflict
           end
         elsif folder
+          # To fullfil Litmus requirements to not delete folder if fragments are in the URL
+          uri = URI(request.get_header('REQUEST_URI'))
+          if uri && uri.fragment
+            raise BadRequest
+          end
           raise Forbidden unless User.current.admin? || User.current.allowed_to?(:folder_manipulation, project)
           raise Forbidden unless DmsfFolder.permissions?(folder, false)
           folder.delete(false) ? NoContent : Conflict
